Add refund address to settings, editable on the account page.
