--- frontend/client/components/Settings/Account.less
@@ -1,20 +1,0 @@
-.AccountSettings {
-  &-form {
-    &-resend {
-      position: relative;
-      margin-bottom: 1rem;
-    }
-
-    & > .ant-form-item {
-      margin-bottom: 0.75rem;
-    }
-
-    & button {
-      margin: 1rem auto 0;
-    }
-  }
-
-  &-alert {
-    margin-top: 1rem;
-  }
-}
