+++ frontend/client/components/Settings/Account/ChangeEmail.less
@@ -1,0 +1,19 @@
+.ChangeEmail {
+  &-resend {
+    position: relative;
+    margin-bottom: 1rem;
+  }
+
+  &-alert {
+    margin-top: 1rem;
+  }
+
+  & > .ant-form-item {
+    margin-bottom: 0.75rem;
+  }
+
+  & button {
+    margin: 1rem auto 0;
+  }
+}
+
